requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.1
pyjwt==2.8.0
jinja2==3.1.2
python-multipart==0.0.6
email-validator
python-dotenv==1.0.0
openai-agents
